#434: Moved the AbilitySkillModifierIncreaseParser and its tests to the model folder
 The file was not modified yet.
