+++ src/app/ffbe/model/effects/abilities/skill/ability-skill-modifier-increase.parser.spec.ts
@@ -1,4 +1,4 @@
-import {AbilityEffectParserFactory} from './ability-effect-parser.factory';
+import {AbilityEffectParserFactory} from '../../../../mappers/effects/abilities/ability-effect-parser.factory';
 import {
   ABILITY_SKILLS_NAMES_TEST_DATA,
   ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA,
@@ -6,11 +6,11 @@ import {
   MAGIC_SKILLS_NAMES_TEST_DATA,
   MAGIC_SKILLS_SHORTDESCRIPTIONS_TEST_DATA,
   MAGIC_SKILLS_TEST_DATA
-} from '../../../model/skill.model.spec';
-import {Skill} from '../../../model/skill.model';
-import {SkillsService} from '../../../services/skills.service';
-import {HTML_LINE_RETURN} from '../skill-effects.mapper';
-import {SkillsServiceMock} from '../../../services/skills.service.spec';
+} from '../../../skill.model.spec';
+import {Skill} from '../../../skill.model';
+import {SkillsService} from '../../../../services/skills.service';
+import {HTML_LINE_RETURN} from '../../../../mappers/effects/skill-effects.mapper';
+import {SkillsServiceMock} from '../../../../services/skills.service.spec';
 
 describe('AbilitySkillModifierIncreaseParser', () => {
